// shared/enums/ClientToSocketMsg.ts
export enum ClientToSocketMsg {
    CONNECTION = "CONNECTION",
    PLAYER_UPDATE = "PLAYER_UPDATE",
    START_ATTACK = "START_ATTACK",
    ATTACK = "ATTACK",
    DISCONNECT = "DISCONNECT",
    SPAWN_PLAYER = "SPAWN_PLAYER",
    RESPAWN_PLAYER = "RESPAWN_PLAYER",
    PLAYER_POS_UPDATE = "PLAYER_POS_UPDATE",
    PLAYER_DIRECTION_UPDATED = "PLAYER_DIRECTION_UPDATED",
}